+++ libs/py/tf/tfvars.py
@@ -111,6 +111,20 @@ def tfvars():
     return TfVars.model_validate_json(content)
 
 
+def env_key(env: Env, tf_vars: TfVars) -> str:
+    """
+    Given an Env obj find its key in TfVars.envs dict
+
+    Args:
+    * env - Env obj
+    * tf_vars - TfVars obj
+
+    Retrurns:
+    * key of found env object
+    """
+    return [k for k, v in tf_vars.envs.items() if v == env][0]
+
+
 def formatted_tfvars():
     tf_vars_dict = tfvars().model_dump()