+++ scripts/init/poststeps.py
@@ -5,7 +5,7 @@ from libs.py.settings import bazel_settings
 
 def clean() -> None:
     os.chdir(bazel_settings.workspace)
-    shutil.rmtree(bazel_settings.tf_template_dir)
+    shutil.rmtree("tests/e2e")
 
 
 if __name__ == "__main__":